Move gather behaviour to Worker class
