+++ Assets/Prefabs/Worker.prefab
@@ -12,9 +12,9 @@ GameObject:
   - component: {fileID: 376293138153766181}
   - component: {fileID: 376293138153766169}
   - component: {fileID: 376293138153766183}
-  - component: {fileID: 376293138153766168}
+  - component: {fileID: 181007309694338732}
   m_Layer: 0
-  m_Name: Unit
+  m_Name: Worker
   m_TagString: Untagged
   m_Icon: {fileID: 0}
   m_NavMeshLayer: 0
@@ -121,7 +121,7 @@ CircleCollider2D:
   m_Offset: {x: 0, y: 0}
   serializedVersion: 2
   m_Radius: 0.4
---- !u!114 &376293138153766168
+--- !u!114 &181007309694338732
 MonoBehaviour:
   m_ObjectHideFlags: 0
   m_CorrespondingSourceObject: {fileID: 0}
@@ -130,14 +130,16 @@ MonoBehaviour:
   m_GameObject: {fileID: 376293138153766180}
   m_Enabled: 1
   m_EditorHideFlags: 0
-  m_Script: {fileID: 11500000, guid: 5188b8c9517ba674d8b245837e6c7fda, type: 3}
+  m_Script: {fileID: 11500000, guid: 5618a45d7e3884754b572ff70b142c2c, type: 3}
   m_Name: 
   m_EditorClassIdentifier: 
   hoverIndicatorPrefab: {fileID: 8694489509828249487, guid: d2890a4cec30af3449391b0d55f96207,
     type: 3}
   selectIndicatorPrefab: {fileID: 1919610854626468484, guid: 7998739ef2cef4242bac97161b1fce17,
     type: 3}
-  moveSpeed: 10
-  softCollisionRadius: 0.6
-  gatherRate: 5
-  maxGatherDist: 1.1
+  softCollisionRadius: 0.5
+  gatherRate: 1
+  maxGatherDist: 2
+  moveSpeed: 3
+  gatherRate: 1
+  maxGatherDist: 1
