+++ Assets/Scripts/Game/Entity/Unit.cs
@@ -2,14 +2,14 @@
 using System.Collections.Generic;
 using UnityEngine;
 
-public class Unit : Selectable
+public abstract class Unit : Selectable
 {
-    [SerializeField] private float moveSpeed = 1f;
     [SerializeField] private float softCollisionRadius = 0.5f;
     [SerializeField] private float gatherRate = 1f;
     [SerializeField] private float maxGatherDist = 2f;
+    [SerializeField] protected float moveSpeed = 1f;
 
-    private UnitState state;
+    protected UnitState state;
     public UnitState State { get { return state; } }
 
     private Vector3 targetPos;
@@ -29,25 +29,11 @@ public class Unit : Selectable
         targetPos = transform.position;
     }
 
-    new void Update()
+    new protected void Update()
     {
-        base.Update();
-
-        // Update indicators
         hoverIndicator.transform.position = transform.position;
         selectIndicator.transform.position = transform.position;
-        UpdateIndicators();
-
-        // Do action based on state
-        switch (state)
-        {
-            case UnitState.RELOCATING:
-                UpdateRelocate();
-                break;
-            case UnitState.GATHERING:
-                UpdateGather();
-                break;
-        }
+        base.Update();
     }
 
     void FixedUpdate()
@@ -72,19 +58,16 @@ public class Unit : Selectable
 
     public void Relocate(Vector3 targetPos)
     {
-        StopGathering();
         this.targetPos = targetPos;
         state = UnitState.RELOCATING;
     }
 
-    public void Gather(Cluster cluster)
+    public virtual void Interact(Vector3 targetPos)
     {
-        assignedCluster = cluster;
-        AssignResource();
-        state = UnitState.GATHERING;
+        Relocate(targetPos);
     }
 
-    private void UpdateRelocate()
+    protected void UpdateRelocate()
     {
         float targetDistSqr = Vector3.SqrMagnitude(transform.position - targetPos);
         if (targetDistSqr < GameManager.UnitSystem.GroupRadiusSqr)
@@ -94,60 +77,6 @@ public class Unit : Selectable
         else
         {
             transform.position = Vector2.MoveTowards(transform.position, targetPos, moveSpeed * Time.deltaTime);
-        }
-    }
-
-    private void UpdateGather()
-    {
-        if (assignedResource == null) return;
-
-        Vector3 resourcePos = assignedResource.transform.position;
-        float resourceDistSqr = Vector3.SqrMagnitude(transform.position - resourcePos);
-        if (resourceDistSqr < maxGatherDist * maxGatherDist)
-        {
-            assignedResource.TakeDamage(gatherRate * Time.deltaTime);
-        }
-        else
-        {
-            transform.position = Vector2.MoveTowards(transform.position, resourcePos, moveSpeed * Time.deltaTime);
-        }
-    }
-
-    private void StopGathering()
-    {
-        if (assignedCluster == null) return;
-
-        assignedResource.RemoveDestroyedListener(HandleResourceDeath);
-        assignedResource = null;
-        assignedCluster = null;
-    }
-
-    private void AssignResource()
-    {
-        float minDistance = float.MaxValue;
-        Resource closestResource = null;
-        foreach (Resource resource in assignedCluster.Resources)
-        {
-            float distanceToNode = Vector3.Distance(resource.transform.position, transform.position);
-            if (minDistance > distanceToNode)
-            {
-                minDistance = distanceToNode;
-                closestResource = resource;
-            }
-        }
-        assignedResource = closestResource;
-        assignedResource.AddDestroyedListener(HandleResourceDeath);
-    }
-
-    private void HandleResourceDeath()
-    {
-        if (assignedCluster != null && assignedCluster.Resources.Count > 0)
-        {
-            AssignResource();
-        }
-        else
-        {
-            StopGathering();
         }
     }
 }
