+++ Assets/Scripts/Game/Entity/Worker.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 5618a45d7e3884754b572ff70b142c2c
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
