+++ Assets/Scripts/Game/System/UnitSystem.cs
@@ -37,14 +37,7 @@ public class UnitSystem : MonoBehaviour
             // Set units destination
             foreach (Unit unit in units)
             {
-                if (GameManager.SelectionSystem.HighlightedCluster != null)
-                {
-                    unit.Gather(GameManager.SelectionSystem.HighlightedCluster);
-                }
-                else
-                {
-                    unit.Relocate(mousePos);
-                }
+                unit.Interact(mousePos);
             }
         }
     }
